refactor: move lending market filters back to market list
