+++ apps/main/src/llamalend/features/market-list/filters/RangeSliderRowFilter.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react'
 import { type DeepKeys } from '@tanstack/table-core'
 import { type FilterProps } from '@ui-kit/shared/ui/DataTable/data-table.utils'
 import { type NumericTextFieldProps } from '@ui-kit/shared/ui/NumericTextField'
-import { type DecimalRangeValue, SliderInput, type SliderInputProps } from '@ui-kit/shared/ui/SliderInput'
+import { type DecimalRangeValue, SliderInput } from '@ui-kit/shared/ui/SliderInput'
 import { Range } from '@ui-kit/types/util'
 import { decimal, formatNumber } from '@ui-kit/utils'
 import { useMaxValue } from './RangeSliderFilter/useMaxValue'
